Automatically select the largest permitted cluster size
Fixing the cluster size at a particular value doesn't work because FAT32
has minimum and maximum values for the number of clusters in the data
area.  Instead, select the largest possible cluster size for the given
filesystem size and use that instead.  (Why use the largest size?  This
increases the likelihood that the cluster size will be a multiple of the
flash page size.)

TODO: Add tests for the cluster size selection code.