Fix reserved sector calculation when aligning to cluster boundaries
When aligning FAT start and size to cluster boundaries, the reserved
sector count must represent an even number of clusters (the eraseblock
size is an even number of clusters, and there are two FATs, so the
difference of the eraseblock size and the space used by FATs must be
even).  Take this into account when computing the maximal FAT size.