Increase FAT32 min clusters
The FAT32 specification allows FAT32 filesystems with 65525 data
clusters, but according to Microsoft, Windows 2000/XP (at least) is
unhappy with fewer than 65527 clusters.  Also, mkdosfs complains when
you try to create a FAT32 filesystem with fewer than 65529 clusters
(though creating the filesystem succeeds, and dosfsck and the Linux FAT
implementation have no problems with it).  A few wasted sectors is a
small price to pay for improved compatibility, so increase our minimum
cluster count to 65529.